Huge Overall
This updateis immense.. besides general refactor to make things cleaner, it also adds browser history functionality to allow going forward and backwards in a given tab just like you would be able to do in a real web browser. I added a refresh button just to refresh the links, because it added to the "browser" theme though it was actually just a work around for the pain of trying to get both colors and links reconstructed from the page memory.